20.06.2023

+++ SHS/src/main/java/com/renewableenergy/SHS/repository/EnergyConsumerRepository.java
@@ -4,7 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.renewableenergy.SHS.entity.EnergyConsumer;
+
+
 @Repository
-public interface EnergyConsumerRepository extends JpaRepository<EnergyConsumer, Integer>{
+public interface EnergyConsumerRepository extends JpaRepository<EnergyConsumer, Long>{
 
 }
