+++ SHS/src/main/java/com/renewableenergy/SHS/repository/EnergyProducerinHomeRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.renewableenergy.SHS.entity.EnergyProducerinHome;
+
 @Repository
-public interface EnergyProducerinHomeRepository extends JpaRepository<EnergyProducerinHome, Integer>{
+public interface EnergyProducerinHomeRepository extends JpaRepository<EnergyProducerinHome, Long>{
 
 }
